html/python/requirements-api.txt
fastapi>=0.128.8
uvicorn[standard]>=0.39.0
python-snap7>=2.0.2
psycopg2-binary>=2.9.10
